Changed recursively Logic

+++ Program.cs
@@ -6,17 +6,36 @@ try
 {
     if (di.Exists)
     {
-        foreach (FileInfo file in di.GetFiles().Where(x => (DateTime.Now - x.LastWriteTime).TotalMinutes > 30))
+        var allFilesPaths = Directory.GetFiles(path, "*.*", SearchOption.AllDirectories);
+        foreach (var filePath in allFilesPaths)
         {
-            file.Delete();
+            if (File.Exists(filePath))
+            {
+                var file = new FileInfo(filePath);
+
+                if ((DateTime.Now - file.LastWriteTime).TotalMinutes > 30)
+                    file.Delete();
+            }
+            else
+                Console.WriteLine($"Некорректный путь файла '{filePath}'!");
         }
-        foreach (DirectoryInfo dir in di.GetDirectories().Where(x => (DateTime.Now - x.LastWriteTime).TotalMinutes > 30))
+
+        var allDirsPaths = Directory.GetDirectories(path, "*.*", SearchOption.AllDirectories);
+        foreach (var dirPath in allDirsPaths)
         {
-            dir.Delete(true);
+            if (Directory.Exists(dirPath))
+            {
+                var dir = new DirectoryInfo(dirPath);
+
+                if ((DateTime.Now - dir.LastWriteTime).TotalMinutes > 30)
+                    dir.Delete();
+            }
+            else
+                Console.WriteLine($"Некорректный путь внутреннего каталога '{dirPath}'!");
         }
     }
     else
-        Console.WriteLine("Некорректный путь!");
+        Console.WriteLine($"Некорректный путь каталога '{path}'!");
 }
 catch (Exception ex)
 {
